requirements.txt
psycopg2-binary
Flask
gunicorn
tqdm

